backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
google-generativeai
pdfplumber
pytesseract
Pillow
